qubo_prover_v3.1/requirements.txt
# QUBO Prover V3.1 - Dependencies
# Core
pyqubo>=1.4.0
dimod>=0.12.0
dwave-neal>=0.6.0

# Optional quantum backends
# openjij>=0.9.0
# dwave-system>=1.18.0

# Neural network
torch>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0

